Set up express server
added functionality for calls to the express server to get infromation using GET request that respond with JSON for the object requested, e.g. if /products is requested it will return the JSON file with all the product info.

Added node modules:
Express - Server
Nodemom - Allows to live update the server without needing to restart it
Concurrently - Allows both front end and back end to be launched with one Command.

Lauching Server:
first you need to cd into the server folder
in the terminal type: npm run dev
this will lauch the react app and the express server at the same time.